+++ backend/src/management/ministries/const/unassigned-member-order.enum.ts
@@ -1,0 +1,3 @@
+export enum UnassignedMemberOrder {
+  REGISTERED_AT = 'registeredAt',
+}
