+++ backend/src/management/ministries/dto/ministry-group/request/member/remove-member-from-ministry-group.dto.ts
